requirements.txt
numpy
fire
transformers>=4.28.1
sentencepiece
tokenizers>=0.13.3
wandb
einops
flask
requests
google-cloud-storage